OE-9300 Make sure secondary diagnosis is found.

+++ protected/modules/OphCiExamination/models/Element_OphCiExamination_Diagnoses.php
@@ -218,6 +218,9 @@ class Element_OphCiExamination_Diagnoses extends \BaseEventTypeElement
         foreach ($current_diagnoses as $cd) {
             if (!array_key_exists($cd->id, $disorder_to_update)) {
                 $secondary_diagnosis = \SecondaryDiagnosis::model()->find('disorder_id = :disorder_id', [':disorder_id' => $cd->disorder_id]);
+                if (!$secondary_diagnosis) {
+                    throw new \Exception("Unable to find secondary disorder linked to disorder $cs->disorder_id");
+                }
                 $this->event->episode->patient->removeDiagnosis($secondary_diagnosis->id);
                 if (!$cd->delete()) {
                     throw new \Exception('Unable to remove old disorder');
